Safari-611.1.10/LayoutTests/fast/scrolling/scroll-to-focused-element-canceled-by-fragment-navigation-expected.txt

This tests that scrolling to the focused element gets overriden by a fragment navigation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.scrollingElement.scrollTop is 0
PASS document.scrollingElement.scrollTop is 0
PASS successfullyParsed is true

TEST COMPLETE
target
